create_job(): Do not create build jobs for arches with pre-existing binaries.
When not creating a build Job, create a Binary instead, so binary jobs can be started immediately.
This requires Binary.build_job to nullable, as there now might be binaries that wasn't built by a debile Job.
As a result Binary.arch must be a column instead of a hybrid_property pointing to the arch of the build job (which may not exist).