Enforce ForeignKey constraints in sqlite, and enable database level ON DELETE triggers.
Note that any ForeignKey with an ondelete="CASCADE" attribute, MUST have a
_reverse_ relationship with a 'cascade' attribute including "delete".

Also note that any ForeignKey in a Column with an nullable=True attribute
MUST NOT use any ondelete attribute value other than "CASCADE" or "SET NULL".

Failure to follow these restrictions will result in a SQLAlchemy session
cache inconsistent with the underlying database, which may lead to heisenbugs.

Reverse relationships of a ForeignKey with an 'ondelete' attribute SHOULD set
passive_deletes=True in order to avoid duplicating the work already done by
the database.